sartopo_python minor fixes
'crop' returned a list of resulting features; 'expand' and 'cut' did not return anything, causing plans console to beep even on success.  Change 'cut' to return a list of resulting features, and 'expand' to return True on success;
in plans console, do an immediate refresh when the Go button is clicked; eventually this will be enhanced to use autocomplete (STSFeatureComboBox as used in sartopo_address).