create and populate new debrief map if requested
SpecifyMapDialog has an added option to show a set of radio buttons to either create a new map or specify an existing map.  Using that dialog still allows the user to create a new map on a different domain and port than the incident map if needed .  Still a few things to clean up - mainly that the new debrief map doesn't show up in the list of maps (on CTD 4214).