Add excluded-folder capability
Hide incident map features from the debrief map by moving them to an incident map folder whose name is in the (currently hardcoded) list of excluded folder titles.

Current list (case insensitive):
aTEAMS
scratch
Tracks to location

We might want to add a GUI option for this list in the future, and a config file setting.