fix sartopo_bg #21
add note in debrief GUI that apptracks will not show up in the debrief sartopo map, but they will show up in generated PDFs.  Not sure if we will find a smarter way to address this in the future.  It is by-design, so that a stale partial apptrack never shows up in the generated debrief map, and the PDF is always generated with the latest available data from the incident map.